Changes instance cards to display metrics charts
