+++ src/client/Components/MetricsTable.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const MetricsTable = ({ instanceId, state, tags, name }) => {
+  return <div>MetricsTable</div>;
+};
+
+export default MetricsTable;
