+++ src/client/styles.css
@@ -2,8 +2,8 @@
 @tailwind components;
 @tailwind utilities;
 
-/* @layer base {
+@layer base {
   html {
     @apply text-discord-100;
   }
-} */
+}
